requirements.txt
nibabel==5.1.0
numpy==1.24.3
pandas==2.0.3
Pillow==9.3.0
Pillow==10.1.0
pydicom==2.3.1
pyodide_http==0.2.1
skimage==0.0
streamlit==1.28.2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
